// typings.d.ts
interface Testimonay {
    name: string;
    position: string;
    image: string;
    quote: string;
}

interface TestimonialsProps {
    testimonay: Testimonay[]
    pageDescription: string;
    pageTitle: string;
    pageSubtitle: string;
    image: string;
}
// ...
interface SliderData {
    img: string,
    title: string,
    description: string,
    button: string,
    time:string,
    location: string,
}